Run SYCL workflows with both nightly and release version of DPC++ (#246)
This will create a job for each version of the compiler (nightly and
release), on both Intel PVC Test and Intel PVC Python Interface Test.

The workflow is still callable with a custom tag of the compiler nightly
build using the `workflow_dispatch` form.

Running Intel PVC tests for the release compiler only works with `icpx`
(not with `clang++`)

---------

Co-authored-by: Alejandro Acosta <alejandro.acosta@codeplay.com>