Fixing the intrinsic  offset for openCL 2d Load/prefetch built-ins functions (#336)
One of the differences between the intrinsic and the built-in function
(the SPIR-V instruction is the same) is that the intrinsic mapped
directly to the HW, so it needed the "-1", but the built-in function and
the SPIR-V instruction should just take the width, height, and pitch
as-is.

When we have the size 1 for the hight, the -1 was causing an issue on
BMG making tests to fail. Since the openCL built-ins mainly used in
prefetch it was causing segfault in the code when prefetch is used.

Co-authored-by: Alejandro Acosta <alejandro.acosta@codeplay.com>