Fix case where matrix size exceeds max uint32 (#364)
This PR fixes a uint32_t overflow issue in the stride offset for copy
operations. I added a test for this via 00_pvc_gemm.

I've also fixed a bug in the RNG used for `initialize_block`: the
generator was being reconstructed each call, which means each thread
would always produce the *same* random number. This was actually masking
the above bug for certain shapes.

---------

Co-authored-by: Alejandro Acosta <alejandro.acosta@codeplay.com>