Avoid warnings about Int (alias template) (#365)
This PR gets rid of these warnings about alias templates:
```
/home/joe/sources/cutlass/include/cute/atom/copy_traits_xe.hpp:413:7: warning: class template argument deduction for alias templates is a C++20 extension [-Wc++20-extensions]
  413 |   Int mma_atom_size_N =
      |       ^
```

Basically `Int` is an alias template:
```cpp
template <int v>
using Int = C<v>;
```
and although the compiler is perfectly capable of determining the
integer template argument to `Int` from e.g.:
```cpp
  Int mma_atom_size_N =
      Int<!TiledCopy::is_convention_MN ? size<1>(mma_atom_shape_2d) : size<0>(mma_atom_shape_2d)>{};
```
it's not permitted under c++17. I chose to solve this by changing to
`auto`. It would be equally valid to change them to `C`:
```cpp
  C mma_atom_size_N = ...
```
or to refactor it like:
```cpp
  Int<!TiledCopy::is_convention_MN ? size<1>(mma_atom_shape_2d) : size<0>(mma_atom_shape_2d)> mma_atom_size_N{}; 
```
but this felt the most readable.

---------

Co-authored-by: Alejandro Acosta <alejandro.acosta@codeplay.com>