Support bfloat16 GEMM accumulation in Python interface (#370)
As a follow-up to
https://github.com/codeplaysoftware/cutlass-sycl/pull/356, support
bfloat16 GEMM accumulation in the Python interface.

The main change is that using bfloat 16 accumulation results in calls to
the bfloat16 device library, which would result in unresolved linkage
when targeting `spir64`. Instead, the compiler now targets
`intel_gpu_pvc`.

A few small tweaks were also necessary to enable argument handling for
bfloat16 in the Python interface.

DPCTL is also now taken directly from the master branch and the
workaround from
https://github.com/codeplaysoftware/cutlass-sycl/pull/242 is dropped as
it was fixed in DPCTL.

`can_implement` was changed to not check the strides of `C` if it's
`nullptr`, which is relevant for the Pytorch integration.

---------

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>